Split up test utils and world_utils
