+++ tests_old/test_format_wrapper/test_encode_decode.py
@@ -3,8 +3,9 @@ import unittest
 
 from amulet_nbt import CompoundTag
 from amulet import load_format
-from data.util import WorldTemp, for_each_world, BaseWorldTest
-from data.worlds_src import levels
+
+from tests.data.world_utils import WorldTemp, for_each_world, BaseWorldTest
+from tests.data.worlds_src import levels
 
 
 @for_each_world(globals(), levels)
